+++ minijinja/src/value.rs
@@ -65,16 +65,12 @@ use std::collections::BTreeMap;
 use std::convert::TryFrom;
 use std::fmt::{self, Write};
 use std::sync::atomic::{self, AtomicBool, AtomicUsize};
-use std::sync::Arc;
 
 use serde::ser::{self, Serialize, Serializer};
 
 use crate::error::{Error, ErrorKind};
 use crate::key::{Key, KeySerializer};
-use crate::utils::matches;
-
-// the type we want to use for refcounting
-pub(crate) type RcType<T> = Arc<T>;
+use crate::utils::{matches, RcType};
 
 // We use in-band signalling to roundtrip some internal values.  This is
 // not ideal but unfortunately there is no better system in serde today.